Remove superfluous nonsense
Since the paradigm shift of both the architecture AND the development
style, a lot of this stuff is either worthless or simply clutters the
space. There is still a lot of cleanup to be done like this, but
requires a bit more finesse because there is still some good logic
floating around, especially in the navigation modules.