Remove sprite/image stuff from the model system
This really has no business being here. The model system does exactly as it was meant to do right now and the addition of visuals to it just adds uneeded complexity to it and it's never even gauranteed that's how a game dev will use it. Let the game dev attach things as they will in their own way.