Rename variable to be meaningful
There is a difference between the actual frames we're given and the logical frames. The logical frames are essentially N (where N is the number directions) frames put into one. So the number of logical frames are the number of frames we have if we only had 1 direction to worry about.