Remove animation related things from the model code
The animations were too restrictive for the game dev because they made tying in-game actions to animations very hard (which will generally be happening a lot).

The wire model now just encapsulates the frame + direction layer and the model is the concrete version of that.

It is expected for the game dev to animate things themselves and to use this as a tool.