Rework input infrastructure to be event-caller rather than poll paradigm
The poll paradigm makes responding to events difficult. It would require the game dev to constantly poll the input object to know what values things are, and would require them to constantly keep track of how things change.

The new way is the common way of doing things and allows the game dev to simply assign handlers so they can fire and forget.