Correct entity rendering
- Add width and height parameters
- Translate entity rendering positions from entity position to draw position
- Correct typo in Direction code

All of these changes are because entity position are considered to be at the bottom center of the entity (specifically fur purposes of world positioning and pathing). Rendering occurs relative to the top-left corner of the frame. These changes are to correctly translate hardpoint positions relative to the entity position.